Change initialization method name and how it's called.
Running this on `init` is silly, as it’s just adding hooks — including
one being added on `init`.

Something that runs on it, adding to init, is kinda silly.  It works,
because it’s added at priority 0, and then adds the others at the
default priority of 10, but better to just add before `init` starts
firing.